Improved the presentation of immediate chunk updates (#3255)
Previously even when meshes were generated in the blocking defer mode "Immediate" (zero frame) they would not become visible if they were part of a previously invisible section and thus not in a render list already. This change makes it so that sections that receive results from blocking rebuilds are added to the render list.

Notably, this fixes a problem where block entities would take one frame to show up after the upload for their containing section has been completed. This change only performs unconditional rendering of sections with newly added block entities if the section was updated with a blocking (immediate) update or if the section is close to the camera.

There was previously an implementation for controlling sort task deferring as part of the sorting code. A different implementation of essentially the same concept but for rebuild tasks was added later. Now they both use the newer approach, which simplifies the task submit code.